fix: change annealingDurationRange to a list of floats. (#100)

+++ src/braket/device_schema/dwave/dwave_2000Q_device_level_parameters_v1.py
@@ -82,7 +82,7 @@ class Dwave2000QDeviceLevelParameters(BraketSchemaBase):
     braketSchemaHeader: BraketSchemaHeader = Field(default=_PROGRAM_HEADER, const=_PROGRAM_HEADER)
     annealingOffsets: Optional[List[float]]
     annealingSchedule: Optional[List[List[float]]]
-    annealingDuration: Optional[int] = Field(gt=0)
+    annealingDuration: Optional[float] = Field(gt=0)
     autoScale: Optional[bool]
     beta: Optional[float]
     chains: Optional[List[List[int]]]
